+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * Add `Container::attach` method
 * Add `Container::changes` method
 * Add `Container::logs` method
+* Add `Container::stats`, `Container::stats_stream`, `Containers::stats` and `Containers::stats_stream` methods